Version 7.2.7 - [Hotfix] Elite Coins are now correctly soulbound - [Fix] All custom bosses are now able to escape instead of just regional bosses
