+++ pom.xml
@@ -7,7 +7,7 @@
 
     <groupId>com.magmaguy</groupId>
     <artifactId>EliteMobs</artifactId>
-    <version>7.2.6</version>
+    <version>7.2.7</version>
     <packaging>jar</packaging>
 
     <properties>
