+++ src/main/java/com/magmaguy/elitemobs/npcs/NPCWorkingHours.java
@@ -2,6 +2,7 @@ package com.magmaguy.elitemobs.npcs;
 
 import com.magmaguy.elitemobs.ChatColorConverter;
 import com.magmaguy.elitemobs.MetadataHandler;
+import com.magmaguy.elitemobs.config.ConfigValues;
 import com.magmaguy.elitemobs.config.TranslationConfig;
 import org.bukkit.potion.PotionEffect;
 import org.bukkit.potion.PotionEffectType;
@@ -32,7 +33,7 @@ public class NPCWorkingHours {
                         npcEntity.getVillager().addPotionEffect(new PotionEffect(PotionEffectType.INVISIBILITY, 12000, 1));
 
                         if (!npcEntity.getIsSleeping()) {
-                            npcEntity.setTempRole(ChatColorConverter.convert(TranslationConfig.NPC_SLEEPING_MESSAGE));
+                            npcEntity.setTempRole(ChatColorConverter.convert(ConfigValues.translationConfig.getString(TranslationConfig.NPC_SLEEPING_MESSAGE)));
                             npcEntity.setIsSleeping(true);
                         }
