+++ src/main/resources/plugin.yml
@@ -1,5 +1,5 @@
 name: EliteMobs
-version: 7.2.6
+version: 7.2.7
 author: MagmaGuy
 main: com.magmaguy.elitemobs.EliteMobs
 api-version: 1.14
